Reset the simulated subsystems when exiting the splash screen
When the demo is idle and goes back to the splash screen it is
convenient to reset the demo when exiting the splash screen to
put the data back to their initial state to allow a new fresh
run through the demo.

This also adds the 's' key to go back to the splash screen.

Issue: eng/projects/sabd/cheri-examples#7